requirements.txt
# Core dependencies
Flask>=2.0.1
Werkzeug>=2.0.1
Jinja2>=3.0.1
itsdangerous>=2.0.1
click>=8.0.1
MarkupSafe>=2.0.1
gunicorn>=20.1.0

# Environment and Config
python-dotenv==1.0.0

# Blog Processing
markdown==3.4.3
python-frontmatter==1.0.0